requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
PyMySQL==1.0.3
toollib==1.5.1
python-dotenv==1.0.1